requirements.txt
langgraph>=0.2.60
langchain-core>=0.3.28
langchain>=0.3.14

langchain-openai>=0.2.14

qdrant-client>=1.12.1

python-dotenv>=1.0.1
pydantic>=2.10.4
pydantic-settings>=2.7.0

psycopg2-binary>=2.9.10
asyncpg>=0.30.0
langgraph-checkpoint-postgres>=2.0.11

tavily-python>=0.5.0

langchain-community>=0.3.14
sentence-transformers>=3.3.1

pytest>=8.3.4
pytest-asyncio>=0.25.0
ruff>=0.8.4
